replace LookupTableWorkflow with GenericUnwrapWorkflow
